# backend/internal/repo/models.go
package repo

import "time"

type User struct {
	ID           int64     `json:"id"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"-"`
	Role         string    `json:"role"`
	CreatedAt    time.Time `json:"created_at"`
}

type Session struct {
	ID        int64     `json:"id"`
	TeacherID int64     `json:"teacher_id"`
	Title     string    `json:"title"`
	StartTime time.Time `json:"start_time"`
	CreatedAt time.Time `json:"created_at"`
}

type HomeworkSubmission struct {
	ID        int64     `json:"id"`
	SessionID int64     `json:"session_id"`
	StudentID int64     `json:"student_id"`
	Content   string    `json:"content"`
	Status    string    `json:"status"`
	CreatedAt time.Time `json:"created_at"`
}

type AuditLog struct {
	ID          int64     `json:"id"`
	ActorUserID int64     `json:"actor_user_id"`
	Action      string    `json:"action"`
	EntityType  string    `json:"entity_type"`
	EntityID    int64     `json:"entity_id"`
	MetaJSON    string    `json:"meta_json"`
	CreatedAt   time.Time `json:"created_at"`
}

type Notification struct {
	ID          int64     `json:"id"`
	EventType   string    `json:"event_type"`
	PayloadJSON string    `json:"payload_json"`
	Status      string    `json:"status"`
	CreatedAt   time.Time `json:"created_at"`
}

type Course struct {
	ID          int64     `json:"id"`
	Title       string    `json:"title"`
	Description string    `json:"description"`
	ImageURL    string    `json:"image_url"`
	CreatedAt   time.Time `json:"created_at"`
}

type Lesson struct {
	ID          int64     `json:"id"`
	CourseID    int64     `json:"course_id"`
	Title       string    `json:"title"`
	Description string    `json:"description"`
	VideoURL    string    `json:"video_url"`
	FileURL     string    `json:"file_url"`
	HLSUrl      string    `json:"hls_url"`
	VideoStatus string    `json:"video_status"`
	SortOrder   int       `json:"sort_order"`
	CreatedAt   time.Time `json:"created_at"`
}

type VideoUpload struct {
	ID               int64     `json:"id"`
	LessonID         int64     `json:"lesson_id"`
	OriginalFilename string    `json:"original_filename"`
	StoredPath       string    `json:"stored_path"`
	HLSPath          string    `json:"hls_path"`
	Status           string    `json:"status"`
	ErrorMessage     string    `json:"error_message"`
	CreatedAt        time.Time `json:"created_at"`
	FinishedAt       time.Time `json:"finished_at"`
}